Obesity {SMIR}	R Documentation

Obesity Data

Description

The obesity data originated from the Muscatine, Iowa study of child development. A component of the Muscatine study was a longitudinal study of childhood obesity, discussed in Woolson, 1984. Part of the data from Woolson and Clarke were reanalysed by Fitzmaurice, 1994 and Aitkin, 1999.

Usage

data(Obesity)

Format

A data frame with 3042 obs. of 9 variables.

a8

a numeric vector

sex

a numeric vector

pattern

a factor w/ 26 levels

age

a numeric vector

status

a numeric vector

id

a numeric vector

id2

a numeric vector

age1

a numeric vector

age2

a numeric vector

Source

Aitkin, M. (1999), "A general maximum likelihood analysis of variance components in generalized linear models", Biometrics, 55, 117–128.

Examples


data(Obesity)

[Package SMIR version 0.01 Index]