Oxboys {SMIR}	R Documentation

Oxford boys height data

Description

The heights of 26 boys measured on nine occasions over 2 years

Usage

data(Oxboys)

Format

A data frame with 234 observations on the following 4 variables.

Subject

a numeric vector

age

a numeric vector

height

a numeric vector

Occasion

a numeric vector

Details

The heights of 26 boys in Oxford were measured on nine occasions over two years, initially at age 12; the occasions were almost equally spaced in time through the school year. The data are the height in cm, age expressed in months relative to an origin of 13 years, and months relative to an annual time origin of January 1.

Source

Harrison, G.A. and Brush, G. (1990). On correlations between adjacent velocities and accelerations in longitudinal growth data. Annals of Human Biology, bf{17(1)}, 55–57.

Examples


data(Oxboys)

[Package SMIR version 0.01 Index]