betablok {SMIR}	R Documentation

Beta-blockers for myocardial infarction

Description

A 22-centre clinical trial of beta-blockers for reducing mortality after myocardial infarction, described by Yusuf et. al. (1984). The important issue is the generalizability of the treatment effect across different patient populations.

Usage

betablok

Format

A dataframe with 44 obs. and 4 variables:

[,1] 	 r 	 integer

[,2] 	 n 	 integer

[,3] 	 centre 	 int

[,4] 	 treat 	 factor w /2 levels "C","T"

Source

Yusuf, S. and Peto, R. and Lewis, J. and Collins, R. and Sleight, P. (1984),"Beta blockade during and after myocardial infarction: an overview of the randomized trials",Progress in Cardiovascular Diseases, 27, 335–371.

[Package SMIR version 0.01 Index]