bronchitis {SMIR}	R Documentation

Chronic bronchitis in a sample of mean in Cardiff

Description

The data consist of observations on three variables for each of 212 men in a sample of Cardiff enumeration districts.

Usage

bronchit

Format

A data.frame of 212 obs of 3 variables:

[,1] 	 r 	 integer, 1= respondent suffered from chronic bronchitis

[,2] 	 cig 	 numeric, the number of cigarettes per day

[,3] 	 poll 	 numeric, the smoke level in the locality

Source

Jones, K. (1975), A geographical contribution to the aetiology of chronic bronchitis, Unpublished BSc dissertation, University of Southampton. Published in Wrigley, N. (1976). Introduction to the use of logit models in geography, Geo.Abstracts Ltd, CATMOG 10, University of East Anglia, Norwich.

[Package SMIR version 0.01 Index]