byssinosis {SMIR}	R Documentation

Byssinosis in the cotton industry

Description

The dataset contains the number of workers in a survey of the US cotton industry suffering and not suffering from the lung disease byssinosis, together with the values of five cross-classifying categorical explanatory variables: the race, sex and smoking habit of the worker, the length of employment in three categories, and the dustiness of the workplace in three categories.

Usage

byssinosis

Format

A data.frame of 72 obs. of 7 variables:

[,1] 	 dust 	 Factor w/ 3 levels "most","less"

[,2] 	 race 	 Factor w/ 2 levels "white","non-white"

[,3] 	 sex 	 Factor w/ 2 levels "male","female"

[,4] 	 smok 	 Factor w/ 2 levels "smoker","non"

[,5] 	 emp 	 Factor w/ 3 levels "<10","10-20",...

[,6] 	 yes 	 int Number of workers who suffered byssinosis

[,7] 	 no 	 int Number of workers who did not suffer from byssinosis

Source

Higgins, J.E. and Koch, G.G. (1977), "Variable selection and generalized chi-square analysis of categorical data to a large cross-sectional occupational health survey", Int. Statist. Rev., 45, 51–62.

[Package SMIR version 0.01 Index]