cars {SMIR}	R Documentation

Performance data for cars from Motor Trend magazine

Description

The data are quarter-mile acceleration time in seconds and fuel consumption in miles per (US) gallon for 32 cars tested by the US Motor Trend magazine in 1974. Nine explanatory variables are given: shape of engine, number of cylinders, transmission type, number of gears, engine displacement in cubic inches, horsepower, number of carburettor barrels, final drive ratio, and weight of the car in thousands of pounds.

Usage

cars

Format

A data.frame of 32 obs. of 11 variables:

[,1] 	 s 	 integer, shape of engine (straight = 1, vee = 0)

[,2] 	 c 	 integer, number of cylinders

[,3] 	 t 	 integer, transmission type, (automatic = 0, manual = 1)

[,4] 	 g 	 integer, number of gears

[,5] 	 disp 	 numeric, engine displacement in cubic inches

[,6] 	 hp 	 integer, horsepower

[,7] 	 cb 	 integer, number of carburettors

[,8] 	 drat 	 numeric, final drive ratio

[,9] 	 wt 	 numeric, weight of car in thousands of pounds

[,10] 	 qmt 	 numeric, quarter-mile acceleration time in seconds

[,11] 	 mpg 	 numeric, fuel consumption in miles per gallon

[,12] 	 model 	 Factor w/ 32 levels "AMC Javelin"...:18 19 ...

Source

Henderson, H.V. and Velleman, P.F. (1981), "Building multiple regression models interactively",Biometrics, 37, 391–411.

[Package SMIR version 0.01 Index]