chd {SMIR}	R Documentation

Coronary heart disease

Description

The file gives the number of men diagnosed as having coronary heart disease (CHD) in an American study of 1329 men (the data are presented and analysed in Ku and Kullback, 1974). The serum cholesterol level and blood pressure in mm mercury were recorded for each man, and are reported in one of four categories, giving a $4 times 4$ cross-classified in each cell of which the number of men with CHD and the total number of men examined are given.

Usage

chd

Format

A data.frame of 16 obs. of 4 variables:

[,1] 	 chol 	 Factor w/ 4 levels "<200","200-219",...

[,2] 	 bp 	 Factor w/ 4 levels "<127","127-146",...

[,3] 	 r 	 integer, number of men with CHD

[,4] 	 t 	 integer, total number of men in study

Source

Ku, H.H. and Kullback, S. (1974), "Loglinear models in contingency table analysis", The American Statistician, 28, 115–22.

[Package SMIR version 0.01 Index]