claims {SMIR}	R Documentation

Insurance claims data

Description

The file gives the number of policyholders of an insurance company who were ``exposed to risk'', and the number of car insurance claims made in the third quarter of 1973 by these policyholders, arranged as a contingency table cross-classified by three four-level factors: the district in which the policyholder lived (1: rural, 2: small towns, 3: large towns, 4: major cities), {tt car}, the engine capacity of the car (1: $<$ 1 litre, 2: 1 – 1.5 litres, 3: 1.5 – 2 litres, 4: $>$ 2 litres), and {tt age}, the age of the policyholder (1: $<$ 25, 2: 25 – 29, 3: 30 – 35, 4: $>$ 35)

Usage

claims

Format

[,1] 	 n 	 integer, number of policy holders

[,2] 	 c 	 integer, number of claims

[,3] 	 age 	 Factor w/ 4 levels "<25","25-29","30-35",">35"

[,4] 	 dist 	 Factor w/ 4 levels "rural","small towns","large towns", "major cities"

[,5] 	 car 	 Factor w/ 4 levels "<1","1-1.5","1.5-2",">2"

Source

Baxter, L.A. and Coutts, S.M. and Ross, G.A.F., 1980, "Application of linear models in motor insurance",Proceedings of the 21st International Congress of Actuaries, Zurich, 11–29.

[Package SMIR version 0.01 Index]