coxph.disparity.R {SMIR}	R Documentation

Check disparity in a Cox Proportional Hazard Model

Description

The coxph.disparity() function returns the disparity from the piecewise exponential model, including all the terms in the likelihood, and is directly comparable to the disparity for the fit of other models used in this chapter.

Usage

coxph.disparity(fit)

Arguments

fit 	 name of an object of class ``coxph''

Details

This form of the likelihood, allows the Cox proportional hazards model to be compared directly to fully parametric models. (Note that log-likelihood value stored in coxph.object is not comparable as it is based on the proportional hazards function and does not include the baseline hazard, this cancels out in the conditional probabilities that form the partial likelihood.)

Value

a num vector

Author(s)

john.hinde@nuigalway.ie

References

Aitkin, M., Francis, B., Hinde, J. and Darnell, R. (2009). Statistical modelling in R, OUP.

Examples


require('SMIR')
require('survival')
data(feigl)
feigl <- within(feigl, {lwbc <- log(wbc)})
feigl.cph <- coxph(Surv(time) ~ ag * lwbc, data = feigl,
                   method = "breslow")
coxph.disparity(feigl.cph)

[Package SMIR version 0.01 Index]