faults {SMIR}	R Documentation

Faults in rolls of material

Description

Bissell gives the numbers of yarn breaks observed in a roll of fabric whilst a textile process was running, as well as the length of the roll of fabric.

Usage

data(faults)

Format

A data.frame of 32 obs. of 2 variables:

[,1] 	 l 	 integer, roll length (m)

[,2] 	 n 	 integer, number of faults

Source

Bissell, A. F. (1972),"A negative binomial model with varying elemental sizes", Biometrika, 59, 435–441.

[Package SMIR version 0.01 Index]