feigl {SMIR}	R Documentation

Leukaemia survival times — Feigl & Zelen

Description

The file contains the survival times in weeks of 33 patients suffering from acute myelogeneous leukaemia, and the values of two explanatory variables, white blood cell count in thousands and a positive or negative factor , positive values being defined by the presence of Auer rods and/or significant granulature of the leukaemic cells in the bone marrow at diagnosis, and negative values if both Auer rods and granulature are absent.

Usage

feigl

Format

A data.frame of 33 obs. of 3 variables:

[,1] 	 wbc 	 numeric, white blood cell count in thousands

[,2] 	 time 	 integer, survival time in weeks

[,3] 	 ag 	 Factor w/ 2 levels "+","-"

Source

Feigl, P. and Zelen, M. (1965). "Estimation of exponential probabilities with concomitant information", Biometrics, 21, 826–38.

[Package SMIR version 0.01 Index]