gehan {SMIR}	R Documentation

Remission times of acute leukemia patients — Gehan et al

Description

Data from a clinical trial which compared 6-mercaptopurine (6-MP) to a placebo in the maintenance of remissions in acute leukemia. The remission times in weeks one year after the start of the study were recorded. Participants were paired according to remission status, an aspect not described in Gehan (1965).

Usage

gehan

Format

A dataframe containing 42 obs. of 5 variables:

[1,] 	 pair 	 numeric defining pair according to remission status

[2,] 	 time 	 numeric time to remission available at the time the trial was stoppped

[3,] 	 cens 	 numeric "0" indicating censored ,"1" uncensored

[4,] 	 treat 	 factor w/ 2 levels "6-MP", "control"

Source

Gehan, E. A. (1965). "A generalized Wilcoxon test for comparing arbitrarily singly-censored samples", Biometrika, 52, 203–233.

References

Freireich, E J et al. (1963)> "The effect of 6-mercaptopurine on the duration of steroid-induced remission in acute leukemia", Gehan, E. A. (1965). "A generalized Wilcoxon test for comparing arbitrarily singly-censored samples", Biometrika, 52, 203–233.

[Package SMIR version 0.01 Index]