ghq {SMIR}	R Documentation

Psychiatric diagnosis based on GHQ

Description

These data were published by Silvapulle, and come from a psychiatric study of the relation between psychiatric diagnosis (as case or non-case) and the value of the score on a 12-item General Health Questionnaire (GHQ), for 120 patients attending a general practitioner's surgery. Each patient was administered the GHQ, resulting in a score between 0 and 12, (however there were no cases or non-cases with GHQ scores of 11 or 12) and was subsequently given a full psychiatric examination by a psychiatrist who did not know the patient's GHQ score. The patient was classified by the psychiatrist as either a ``case'', requiring psychiatric treatment, or a ``non-case''.

Usage

ghq

Format

[,1] 	 sex 	 Factor w/ 2 levels "men","women"

[,2] 	 ghq 	 integer, score from 0,...,12

[,3] 	 c 	 integer, number of patients considered a "case"

[,4] 	 nc 	 integer, number of patients considered a "non-case"

Source

Silvapulle, M. J. (1981), "On the existence of maximum likelihood estimators for the binomial response model", J. Roy. Statist. Soc. B., bf{43}, 310–13.

[Package SMIR version 0.01 Index]