hostility {SMIR}	R Documentation

Bennett's hostility data

Description

A measure of hostility based on word use exhibiting hostility by husbands of wives who had been admitted to hospital after suicide attempts by taking drug overdoses compared to a ``control'' group of husbands.

Usage

data(hostility)

Format

A data frame with 67 observations on the following 10 variables.

group

a numeric vector

nationality

a numeric vector

po

a factor with levels none previous

in.host

a numeric vector

amb.host

a numeric vector

out.host

a numeric vector

covert.host

a numeric vector

positivity

a numeric vector

g

a factor with levels overdoses F controls T controls

nation

a factor with levels Australian British

Source

Bennett, M. D. (1974). The emotional response of husbands to suicide attempts by their wives, Sydney University, Unpublished MD thesis.

Examples


data(hostility)
## maybe str(hostility)
plot(hostility)

[Package SMIR version 0.01 Index]