insult {SMIR}	R Documentation

Effects of unprovoked verbal attack

Description

Ten male and nine female subjects were asked to fill out a questionnaire which mixed innocuous questions with questions attempting to assess the subject's self-reported hostility. A hostility score for each individual was calculated from these responses. After completing the questionnaire, the subjects were then left waiting for a long time, and were subjected to insults and verbal abuse by the experimenter when the questionnaire was eventually collected. All subjects were told that they had filled out the questionnaire incorrectly, and were instructed to fill it out again. A second hostility score was then calculated from these later responses.

Usage

insult

Format

A data.frame of 19 obs. of 3 variables:

[,1] 	 hbefore 	 integer, hostility score before verbal attack

[,2] 	 hafter 	 integer, hostility score after verbal attack

[,3] 	 sex 	 Factor w/ 2 levels "female","male"

Source

Erickson, B. H. and Nosanchuk, T. A., (1979). Understanding Data, Milton Keynes, UK, Open University Press, UK.

[Package SMIR version 0.01 Index]