miners {SMIR}	R Documentation

Pneumoconiosis in coal miners

Description

The file gives the numbers of coalminers classified by radiological examination into one of three categories of pneumoconiosis, normal, mild pneumoconiosis and severe pneumoconiosis, and by years spent working at the coalface (interval midpoint).

Usage

miners

Format

A data.frame of 8 obs. of 4 variables:

[,] 	 years 	 numeric, years (midpoint) of years spent at coalface

[,] 	 n 	 integer, number of miners classified as normal

[,] 	 m 	 integer, number of miners with mild pneumoconiosis

[,] 	 s 	 integer, number of miners with severe pneumoconiosis

Source

Ashford, J. R., (1959). "An approach to the analysis of data from semi-quantal responses in biological response",Biometrics,15,573–581.

[Package SMIR version 0.01 Index]