prentice {SMIR}	R Documentation

Veteran's Administration Lung Cancer Trial

Description

The file consists of survival times in days of 137 lung cancer patients from a Veteran's Administration Lung Cancer trial, together with explanatory variables: performance status, a measure of general medical status on a continuous scale 1–9.9, with 1–3 completely hospitalized, 4–6 partial confinement to hospital, 7–9.9 able to care for self; age in years; time in months from diagnosis to starting on the study; a factor prior therapy (1 no, 2 yes); a factor treatment (1 standard, 2 test) and a factor tumour type (1 squamous, 2 small, 3 adeno, 4 large). There are three censored observations.

Usage

prentice

Format

A data.frame of 137 obs. of 8 variables:

[,1] 	 treat 	 integer, 1= standard, 2= test

[,2] 	 type 	 integer, 1= squamous, 2 =small, 3= adeno, 4= large

[,3] 	 time 	 integer, survival time in days

[,4] 	 censor 	 integer, censoring indicator

[,5] 	 status 	 integer, general medical status on a scale 1–9.9

[,6] 	 mfd 	 integer, time in months from diagnosis

[,7] 	 age 	 integer, age in years

[,8] 	 prior 	 integer, prior therapy 1=no, 2=yes

Source

Prentice, R. L. (1973), "Exponential survivals with censoring and explanatory variables", Biometrika, 60, 279–88.

[Package SMIR version 0.01 Index]