solv {SMIR}	R Documentation

Children's block design test

Description

A sample of twenty-four children was randomly drawn from the population of fifth-grade children attending a state primary school in a Sydney suburb. Each child was assigned to one of two experimental groups, and given instructions by the experimenter on how to construct, from nine differently coloured blocks, one of the $3 times 3$ square designs in the Block Design subtest of the Wechsler Intelligence Scale for Children (WISC). Children in the first group were told to construct the design by starting with a row of three blocks (row group), and those in the second group were told to start with a corner of three blocks (corner group). The total time in seconds to construct four different designs was then measured for each child.

Before the experiment began, the extent of each child's ``field dependence'' was tested by the Embedded Figures Test (EFT), which measures the extent to which subjects can abstract the essential logical structure of a problem from its context (high scores corresponding to high field dependence and low ability).

Usage

solv

Format

A data.frame of 24 obs. of 4 variables:

[,1] 	 child 	 integer, child id

[,2] 	 group 	 Factor w/ 2 levels "corner","row"

[,3] 	 time 	 integer, time in seconds

[,4] 	 eft 	 integer, EFT score

Source

Aitkin, M. and Francis, B. and Hinde, J. (2005), Statistical Modelling in GLIM 4, Oxford University Press, UK

[Package SMIR version 0.01 Index]