stackloss {SMIR}	R Documentation

The Brownlee stackloss data

Description

The data consist of 21 observations on stack-loss (the loss of acid through the stack) in a chemical plant for the conversion of ammonia to nitric acid, with three explanatory variables: air flow, cooling water inlet temperature and acid concentration.

Usage

stackloss

Format

A data.frame of 21 obs. of 4 variables:

[,1] 	 y 	 integer, loss of acid through the stack

[,2] 	 x1 	 integer, air flow

[,3] 	 x2 	 integer, cooling water inlet temperature

[,4] 	 x3 	 integer, acid concentration

Source

Lange, K L and Little, R J A and Taylor, J M G, (1989). "Robust statistical modeling using the $t$ distribution", J Amer Statist Assoc,84, 881–896.

[Package SMIR version 0.01 Index]