statlab {SMIR}	R Documentation

STATLAB census data

Description

The STATLAB Census covers 1296 member families of the Kaiser Foundation Health Plan (a prepaid medical care program) living in the San Francisco Bay Area during the years 1961 - 1972. These families were participating members of the Child Health and Development Study conceived and directed by Jacob Yerushalmy, for many years Professor of Biostatistics in the School of Public Health, University of California, Berkeley.

On her first visit to the Oakland hospital of the Health Plan after pregnancy was diagnosed, each woman was interviewed intensively on a wide range of medical and socioeconomic matters relating both to herself and to her husband. In addition, various physical and physiological measures were made. When her child was born, further data about her and her newborn baby were recorded. Approximately 10 years later the child and mother were called in for follow-up testing, interviewing, and measurement. In some instances, the husband was also interviewed and measured.

The 1296 families of the STATLAB Census are divided into two equal subpopulations: 648 families consisting of a mother, father, and female child; and 648 families of a mother, father, and male child. The children were all born in the Kaiser Foundation Hospital, Oakland, California, between 1 April 1961 and 15 April 1963. The Census does not cover any other children who may also have existed in these families.

Usage

statlab

Format

A data.frame of 1296 obs. of 34 variables:

[,] 	 id 	 integer,

[,] 	 c.b.blood 	 Factor w/ 9 levels

[,] 	 c.b.lgth 	 numeric,

[,] 	 c.b.wgt 	 numeric,

[,] 	 c.b.mo 	 integer,

[,] 	 c.b.day 	 integer,

[,] 	 c.b.hour 	 integer,

[,] 	 c.t.hght 	 numeric,

[,] 	 c.t.wgt 	 integer,

[,] 	 c.t.l 	 Factor w/ 8 levels

[,] 	 c.t.pea 	 integer,

[,] 	 c.t.ra 	 integer,

[,] 	 m.b.blood 	 Factor w/ 9 levels

[,] 	 m.b.ag 	 integer,

[,] 	 m.b.wgt 	 integer,

[,] 	 m.b.o 	 Factor w/ 8 levels

[,] 	 m.b.sm 	 Factor w/ 31 levels

[,] 	 m.t.hght 	 numeric,

[,] 	 m.t.wgt 	 integer,

[,] 	 m.t.e 	 Factor w/ 5 levels

[,] 	 m.t.o 	 Factor w/ 8 levels

[,] 	 m.t.sm 	 Factor w/ 26 levels

[,] 	 f.b.ag 	 integer,

[,] 	 f.b.o 	 Factor w/ 9 levels

[,] 	 f.b.sm 	 Factor w/ 32 levels

[,] 	 f.t.hght 	 numeric,

[,] 	 f.t.wgt 	 integer,

[,] 	 f.t.e 	 Factor w/ 5 levels

[,] 	 f.t.o 	 Factor w/ 9 levels

[,] 	 f.t.sm 	 Factor w/ 32 levels

[,] 	 family.i.b 	 integer,

[,] 	 family.i.t 	 integer,

[,] 	 family.c 	 Factor w/ 6 levels

[,] 	 sex 	 Factor w/ 2 levels "boy","girl"

Source

Hodges, J.L. Krech, D. and Crutchfield, R.S. (1975). StatLab: An Empirical Introduction to Statistics, McGraw-Hill Ryerson, Toronto

[Package SMIR version 0.01 Index]