toxoplas {SMIR}	R Documentation

Toxoplasmosis in El Salvador

Description

The file shows the number of men tested and the number with a positive test for toxoplasmosis in 34 cities in El Slavador, together with the annual rainfall in metres.

Usage

toxoplas

Format

A data.frame of 34 obs. of 3 variables:

[,1] 	 y 	 integer, number of men with a positive test

[,2] 	 n 	 integer, number of men tested

[,3] 	 x 	 numeric, annual rainfall in metres,

Source

Efron, B., (1986),Double exponential families and their use in generalized linear regression, J Amer Statist Assoc,81, 709–721.

[Package SMIR version 0.01 Index]