treg {SMIR}	R Documentation

t-regression model fit

Description

Robust regression by modelling errors as t-distributed with known degrees of freedom rather than normal

Usage

treg(lm.object, r, verbose=TRUE)

Arguments

lm.object 	 An object of class "lm"

r 	 a vector of degrees of freedom

verbose 	 TRUE prints estimates for -2 log likelihood, sigma, and r at each interation.

Details

Fits the $t$ distribution for known degrees of freedom , $r$, and computes the profile likelihood and obtains the joint MLEs of the regression coefficients, sigma and disparity of a robust regression.

Value

an object of class ``treg''

weights 	 working weights

disparity 	 disparity, i.e. full likelihood

tcoef 	 robust regression parameter estimates

r 	 degrees of freedom

sigma 	 estimate of residual standard deviation

Author(s)

ross.darnell@csiro.au

References

Aitkin, M., Francis, B., Hinde, J. and Darnell, R. (2008). Statistical modelling in R, OUP.

See Also

SMIR::summary.treg

Examples


library(SMIR)
data(stackloss)
stackloss.lm <- lm(y ~ x1 + x2 + x3, data = stackloss)
(stackloss.treg1.1 <- treg(stackloss.lm , r=1.1, verbose = FALSE) )

[Package SMIR version 0.01 Index]