trypanos {SMIR}	R Documentation

Trypanosome death data

Description

The file shows the numbers of trypanosomes killed out of a number treated a different dose levels.

Usage

toxoplas

Format

A data.frame of 34 obs. of 3 variables:

[,1] 	 y 	 integer, number of trypanosomes killed

[,2] 	 n 	 integer, number of trypanosomes treated

[,3] 	 x 	 numeric, treatment dose level

Source

Follman, D.A. and Lambert, D. (1989), Generalizing logistic regression by nonparametric mixing, J Amer Statist Assoc, 84, 295–300.

[Package SMIR version 0.01 Index]