vaso {SMIR}	R Documentation

Gilliatt's vaso-constriction data

Description

These data were obtained in a carefully controlled study of the effect of the rate and volume of air inspired by human subjects on the occurrence or non-occurrence of a transient vasoconstriction response in the skin of the fingers.

Format

A data.frame of 39 obs. of 4 variables:

[,1] 	 Subject 	 Factor w/ 3 levels "1","2","3", ...

[,2] 	 Volume 	 numeric, volume of air expired

[,3] 	 Rate 	 numeric, rate of air expired

[,4] 	 Y 	 integer, transient vasoconstriction response, 1=yes, 0=no

Source

Finney, D.J. (1947)., "The estimation from individual records of the relationship between dose and quantal response", Biometrika, 34, 320-34.

References

Gilliatt, R.W. (1948). "Vaso-constriction in the finger after deep inspiration", J. Physiol., 107, 76–88.

[Package SMIR version 0.01 Index]