vietnam {SMIR}	R Documentation

University of North Carolina Vietnam War Student Survey

Description

A survey of student opinion on the Vietnam War was taken at the University of North Carolina in Chapel Hill in May 1967 and published in the student newspaper. Students were asked to fill in ``ballot papers'', available in the Student Council building, stating which policy out of A, B, C or D they supported. Responses were cross-classified by sex and by undergraduate year or graduate status. The policies were:


A:   The US should defeat the power of North Vietnam by 
     widespread bombing of its industries, ports and 
     harbours and by land invasion.

B:   The US should follow the present policy in Vietnam.

C:   The US should de-escalate its military activity, 
     stop bombing North Vietnam, and intensify its efforts 
     to begin negotiation.

D:   The US should withdraw its military forces from Vietnam 
     immediately.

Usage

vietnam

Format

A data.frame of 40 obs. of 4 variables:

[,1] 	 policy 	 Factor w/ 4 levels "A","B","C","D"

[,2] 	 year 	 Factor w/ 5 levels "1","2","3","4","5"

[,3] 	 sex 	 Factor w/ 2 levels "female","male"

[,4] 	 count 	 integer , the number of students in each cell

Source

Aitkin, M. (1996). "A short history of a Vietnam War attitude survey.", Stats, 17, 1–9.

[Package SMIR version 0.01 Index]