woolson {SMIR}	R Documentation

Woolson and Clarke's Obesity Study

Description

This is a subset of the Obesity dataset. Binary indicators of obesity on 1014 children who were 7-9 years old in 1977, and were followed up in 1979 and 1981. Children were classified as obese their weights were more than 210% of the population median weight for their gender and height.

Usage

data(woolson)

Format

A data frame with 48 observations on the following 4 variables.

x

a numeric vector

y

a numeric vector

sex

a numeric vector

age

a numeric vector

Source

Woolson, R.F. and Clark, W.R.(1984). Analysis of categorical incomplete longitudinal data. JRSS A. bf{147}, 87–99.

Examples


data(woolson)

[Package SMIR version 0.01 Index]